fix(security): delete grant once used (#68910)
When an `ApiToken` is created from an `ApiGrant` we should delete that
specific grant from the database in order prevent reuse of the same code
for token exchanges.

Note, we already do this in the "grant exchanger". This particular code
path does not use the generic `GrantExchanger`. I'm not sure why, but
that's something we can come back to refactor.

https://github.com/getsentry/sentry/blob/80f9de5890cac1a7d70252f7f4818baaf2155dd9/src/sentry/mediators/token_exchange/grant_exchanger.py#L37-L39