feat(trace-explorer): Load more spans results per trace (#69305)
The fast query to narrow down a list of trace ids often only returns 1
span per trace. This splits the span fetching into 2 stages. 1 to fetch
trace ids fast, then another to fetch the spans in the trace that match
the conditions. It also includes a suggested spans list that can be used
to render some suggestions based on a different set of conditions.