chore(hc): Disable AI suggested solution in DE region (#69275)
This is a temporary fix to disable AI suggested solution in the DE
region. A separate PR will update organization settings to disable the
option for regions which do not support AI suggested solution. This will
all be eventually replaced with Autofix.

We'll incorporate AI suggested solution into the same changes we're
making in #68315