fix(setup-wizard): Always create a new user API token (#69388)
In the wizard endpoint, we’d reuse existing user auth tokens of the
authenticated user if:
1. the user was part of multiple orgs (==> we can't create an org-based
token)
2. AND we found one that satisfied the necessary permissions for
sourcemap upload.
 
With https://github.com/getsentry/sentry/pull/68148 being merged, we
cannot do this anymore. Plain user auth token values are only gonna be
available directly after the token was created.

For the fix, this PR makes a change to the wizard endpoint to always
create a new user API token. This now works just like when we create an
org token for single-org users.

Closes: https://github.com/getsentry/sentry/pull/69381

---------

Co-authored-by: Daniel Griesser <daniel.griesser.86@gmail.com>