feat(issue-platform): adds a partition key to issue platform producer (#69431)
This PR adds a setting called `issue_platform.use_kafka_partition_key`
that when enabled will set a partition key when we produce to the issue
platform consumer. We partition by the first fingerprint that we find.
Note that technically there could be multiple fingerprints but I believe
we can just use the first fingerprint.