fix(query): Guard against ordering by non unique columns in `RangeQuerySetWrapper` (#69367)
Follow-up from inc-709. This incident was caused by
`RangeQuerySetWrapper` looping infinitely and slamming queries into the
database. It did this because we ordered on a non-unique column, and
currently the logic here will loop forever when the last two rows of the
result set have the same value for the sort col.