fix(routes): Use different route to perform event redirect (#69439)
There are some org slugs that collide with marketing routes. When this
happens, this request to attempt a redirect gets routed to vercel and
results in an unexpected 404. This same redirect endpoint is also
available at a different route that looks like
`/organizations/{org_slug}/projects{proj_slug}/events/{event_id}` that
will not be redirected. So change the endpoint to avoid the issue with
the redirect.