fix(backup): Apply prod-UserOption fix to comparison (#69460)
The fixes in https://github.com/getsentry/sentry/pull/68943 and
https://github.com/getsentry/sentry/pull/69027 worked, but now fail on
comparison. This PR makes comparison clever enough to detect these
"duplicate" UserOption entries and avoid them.

This is a temporary fix. A more robust solution will prevent ever
exporting such UserOptions, but that will take a while to percolate
through the support window, so this should hold us over for the time
being.