feat(uptime): Hook up subscription creation to the detector
This allows the detector to start creating subscriptions for valid urls. It also adds a mode column to the `ProjectUptimeSubscription` table. We put it on this table and not the `UptimeSubscription` table since we want each project to perform validation itself, and also because we don't necessarily know that the subscription has been validated before - possibly it's a manual subscription that a user set up.

This isn't fully complete but is directional. Should mostly need tests to validate behaviour