fix(snuba): Allow common suffixes when validating referrers (#73734)
The query builder adds `".primary"` and `".secondary"` to the query
referrer string when it constructs compound queries (when fetching from
multiple tables). It's a pain to have to declare both a primary _and_ a
secondary _and_ a regular referrer for each of those cases, especially
because the fields in the query can change for all kinds of reasons.
Instead, add a list of common referrer suffixes that are considered
valid. Closes https://github.com/getsentry/sentry/issues/71570