chore(slack): Fix Bug when Slack Rate Limits in `get_channel_id` (#73737)
There was a small regression in `get_channel_id`. When Slack Rate
Limits, the previous implementation forced the caller handles the
exception, in the rewrite, we weren't raising it so I fixed it.

Previous logic:

https://github.com/getsentry/sentry/blob/master/src/sentry/integrations/slack/utils/channel.py#L323-L328

Surprised we haven't run into this, but this probably cause people
aren't abusing alerts creation.

I also made the tests reflect the Rate Limiting logic.