fix(useDimensions): Block render until dimensions are recorded (#73958)
Before this change, useDimensions  was rendering with 0
height/width before the `onResize` kicked in and updated the dimensions.
It looks like
[useResizeObserver](https://github.com/adobe/react-spectrum/blob/main/packages/@react-aria/utils/src/useResizeObserver.ts)
uses `useEffect` which is why we see this. I've added a
`useLayoutEffect` on first render to ensure that the dimensions have
been recorded first before painting. There is a slight performance hit
here, but since the resizes are still in a useEffect, this only impacts
the first render.