fix: lint `.ts` and `.js` files in CI (#73949)
We've been seeing issues lately where bad formatting in `.ts` files is
making it into `master` and blocking deploys. This PR fixes the issue in
`sentry`, `getsentry` PR coming separately but will look the same.

The flow is:

- someone makes a `.ts` change in an editor that doesn't run Prettier
- local `pre-commit` doesn't run Prettier on `.ts` files (Problem 1)
- the `pre-commit` PR check doesn't run Prettier on `.ts` files either,
so there's no auto-fix
- the `prettier (changed files only)` task on the PR doesn't see any
changed files, and doesn't run so there's no PR failure (Problem 2)
- the PR is merged into `master`. `typescript-and-lint` runs on _all_
files and finds the bad formatting

Problem 1: This PR updates the `pre-commit` config which will auto-lint
the files before they're committed (in most cases) and make sure that
`pre-commit` auto-fixes the files in the PR (in all cases)

Problem 2: This PR fixes a typo in the file names that prevented those
names from being passed through to the tasks. With this fix, the
`typescript-and-lint