fix(insights): Span sample panel search clobbers global filter selection (#74013)
The reported bug is that entering a query into the span sample panel
search bar clobbers the global date and environment selection. The
reason is that the `onSearch` handler of the search bar was omitting
some URL parameters.

Always unsplat `location.query` into the URL. `location.query` has _all_
the URL query parameters. `query` only has a subset, depending on what
was passed to `useLocationQuery`