feat(query-builder): Click and drag to select tokens (#73959)
We already have the ability to `cmd + a` to select all, but this adds
click and drag selection similar to a native textarea.

The bulk of the logic is contained in the `useSelectOnDrag` hook, which
watches on mouse events to change the selection. Some changes in other
files were necessary to get partial selection working.