feat(mep): Add fallback to transactions (#73893)
To split the discover dataset in dashboards, we need to be able to run a
query against transactions as the fallback for metrics enhanced queries
instead of the discover dataset.

This adds a branch of code that will allow us to handle transaction-like
requests when we cannot fulfill the query with metrics. Currently it
isn't being used, but will be used when we need to remove our dependency
on discover.