feat(feedback-form): Add a global form instance and hook for opening feedback forms (#74098)
The current way we open feedback forms is by using `useFeedbackWidget()`
which requires a button ref to attach listeners and removes the form
component when the connected button unmounts (this is necessary because
we need to clean up the created form elements on the page somehow).

This PR adds a top-level context and form context that can be used
throughout the app to open forms. This avoids the need to clean up for
every usage and allows the form to remain open even if the trigger no
longer exists.