ref: Remove calls to process_event_proguard (#74075)
This is a partial revert of
https://github.com/getsentry/sentry/pull/64946. In that PR, we
introduced a separate proguard queue to stop slow proguard events from
bogging down the whole processing pipeline. The proguad situation is
much improved, so hopefully we won't need this for much longer.

Turning the option down to 10% had no discernible effect (as expected,
since the process-event-proguard workers have nothing to do anyway). I
think it is safe to remove this logic.

Follow-ups:
* Remove the task
* Remove the queue
* Remove the option