feat(uptime): Add `mode` to `ProjectUptimeSubscription` (#74141)
This adds `mode` to `ProjectUptimeSubscription`. Mode represents the
current way that the subscription is running. MANUAL is reserved for the
future and is for when users explicitly create an uptime monitor
AUTO_DETECTED_ONBOARDING is used when we detect a url automatically and
want to verify that it's a stable url that is worth monitoring.
AUTO_DETECTED_ACTIVE is used once we've decided that a url is stable and
we want to monitor it on an ongoing basis.

Also updates the constraints to incorporate the field.