feat(profiling) render continuous profile chunk flamechart (#74135)
Renders continuous flamechart for a profile chunk.

Disclaimer:
This is one big copy paste of the flamechart component with removal of
span tree code and any reference to the old txn model. Since the
underlying components are not tied to this model anymore, it means they
can be wired up to work with the continuous model. I opted for this to
have cleaner separation and make it easier to make modifications without
having to think of both models - since the main component mostly just
wires up the different parts of the UI, I think that is a good approach
to take