feat(trace-view): Add UI fallback to old view for STs (#73684)
### Summary
Singletenant (and on-prem etc.) may not be fully rolled out for span
extraction. This means the new trace-view which is entirely span based
won't work. This causes links to fallback to the old trace view if the
organization is one without span-extraction rolled out, then proceeds to
check the rollout date so we can offer the new experience where possible
right away.


#### Aside
- `OrganizationSummary` didn't need to be narrowed, so switched it out
for the full `Organization` provided at the top level by
`useOrganization`, can't see a reason why we wouldn't.