feat(query-builder): Allow comparison operators for release keys (#74362)
The search config already calls out these keys as ones which should
allow extra operators.

I've decided to add this to the field definition so that we can both
pass it to the search parser config _and_ use it in the searchbar
component. We can remove these default config entries once the
SmartSearchBar is fully removed.