feat(issue-details): Add new search component to all events tab (#74506)
Under the `issue-stream-search-query-builder` flag, replaces the old
SmartSearchBar component with the new one.

To load the available tags, it uses the same group tags endpoint that
loads the tags summary on issue details. This is an improvement over
what existed prior, since it was loading all available organization tags
event if they did not exist in the issue you're looking at.