fix(query-builder): Better support for adding filter keys by typing (#74817)
When selecting a filter key through the menu, we add a default value in
the `replaceFocusedWordWithFilter()` function. When typing (`<filter>:`)
we don't do that, which results in a filter without a value. This wasn't
as apparent before because we weren't highlighting invalid filters.

These changes will ensure that this function gets called when we detect
a filter after `:` is typed.

Also added an analytics event so that we can track how common this is.